updating the JOIN and USER and PASS and NICK and MODE work as expected

+++ src/main.cpp
@@ -12,7 +12,7 @@ int main(int arg_num, char **args)
 	try{
 		signal(SIGINT, Server::SignalHandler); //-> catch the signal (ctrl + c)
 		signal(SIGQUIT, Server::SignalHandler); //-> catch the signal (ctrl + \)
-		ser.ServerInit(std::stoll(args[1]), args[2]); //-> initialize the server
+		ser.ServerInit(static_cast<int>(std::stoll(args[1])), args[2]); //-> initialize the server
 	}
 	catch(const std::exception& e){
 		ser.CloseFds(); //-> close the file descriptors
